requirements.txt
websocket-client
stomper
